// app/enums/univ-calendar-indexes.enum.ts
    // Indexes of the elements of the calendar data from Univ
    export enum UnivCalendarLineIndexesEnum {
        DTSTAMP = 0,
        DTSTART = 1,
        DTEND = 2,
        SUMMARY = 3,
        LOCATION = 4,
        DESCRIPTION = 5,
        UID = 6,
        CREATED = 7,
        LAST_MODIFIED = 8,
        SEQUENCE = 9
    }